requirements.txt
streamlit==1.35.0
snowflake-connector-python==3.10.0
pandas==2.2.2